Upscaling images with grunt
Fixed some issues with accesibility putting the right roles to each html section

Made the reviews list a ul>li as it's in fact a list

Fixed some styling

I've changed the DBHelper so it's an object and doesn't have static methods. After that, I can 'cache' the json response storing it in a property of the DBHelper so it won't be requested multiple times. This is a huge performance improvement, so the restaurants.json will be only requested once, but everything will work as it was